Extract CMS geography source parser
